+++ src/mock/WETH9.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.12;
 import "@openzeppelin-upgrades/contracts/token/ERC20/ERC20Upgradeable.sol";
 
 contract WETH9 is ERC20Upgradeable {
-    event Deposit(address indexed dst, uint wad);
-    event Withdrawal(address indexed src, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
     function initialize() public initializer {
         __ERC20_init("Wrapped ETH", "WETH");
@@ -20,7 +20,7 @@ contract WETH9 is ERC20Upgradeable {
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         require(balanceOf(msg.sender) >= wad);
         _burn(msg.sender, wad);
         payable(msg.sender).transfer(wad);
