+++ src/protocol/LagrangeCommittee.sol
@@ -10,12 +10,7 @@ import "../interfaces/ILagrangeCommittee.sol";
 import "../interfaces/ILagrangeService.sol";
 import "../interfaces/IVoteWeigher.sol";
 
-contract LagrangeCommittee is
-    Initializable,
-    OwnableUpgradeable,
-    HermezHelpers,
-    ILagrangeCommittee
-{
+contract LagrangeCommittee is Initializable, OwnableUpgradeable, HermezHelpers, ILagrangeCommittee {
     uint8 public constant UPDATE_TYPE_REGISTER = 1;
     uint8 public constant UPDATE_TYPE_AMOUNT_CHANGE = 2;
     uint8 public constant UPDATE_TYPE_UNREGISTER = 3;
@@ -50,28 +45,18 @@ contract LagrangeCommittee is
     mapping(uint32 => uint256) public updatedEpoch;
 
     // Event fired on initialization of a new committee
-    event InitCommittee(
-        uint256 chainID,
-        uint256 duration,
-        uint256 freezeDuration
-    );
+    event InitCommittee(uint256 chainID, uint256 duration, uint256 freezeDuration);
 
     // Fired on successful rotation of committee
     event UpdateCommittee(uint256 chainID, bytes32 current);
 
     modifier onlyService() {
-        require(
-            msg.sender == address(service),
-            "Only Lagrange service can call this function."
-        );
+        require(msg.sender == address(service), "Only Lagrange service can call this function.");
         _;
     }
 
     modifier onlyServiceManager() {
-        require(
-            msg.sender == voteWeigher.serviceManager(),
-            "Only Lagrange service manager can call this function."
-        );
+        require(msg.sender == voteWeigher.serviceManager(), "Only Lagrange service manager can call this function.");
         _;
     }
 
@@ -103,21 +88,10 @@ contract LagrangeCommittee is
     }
 
     // Initialize new committee.
-    function _initCommittee(
-        uint32 chainID,
-        uint256 _duration,
-        uint256 freezeDuration
-    ) internal {
-        require(
-            committeeParams[chainID].startBlock == 0,
-            "Committee has already been initialized."
-        );
+    function _initCommittee(uint32 chainID, uint256 _duration, uint256 freezeDuration) internal {
+        require(committeeParams[chainID].startBlock == 0, "Committee has already been initialized.");
 
-        committeeParams[chainID] = CommitteeDef(
-            block.number,
-            _duration,
-            freezeDuration
-        );
+        committeeParams[chainID] = CommitteeDef(block.number, _duration, freezeDuration);
         committees[chainID][0] = CommitteeData(0, 0, 0);
         committeeLeaves[chainID] = new uint256[](0);
 
@@ -128,14 +102,9 @@ contract LagrangeCommittee is
         return operators[operator].serveUntilBlock;
     }
 
-    function updateOperator(
-        OperatorUpdate memory opUpdate
-    ) external onlyServiceManager {
+    function updateOperator(OperatorUpdate memory opUpdate) external onlyServiceManager {
         if (opUpdate.updateType == UPDATE_TYPE_AMOUNT_CHANGE) {
-            operators[opUpdate.operator].amount = voteWeigher.weightOfOperator(
-                opUpdate.operator,
-                1
-            );
+            operators[opUpdate.operator].amount = voteWeigher.weightOfOperator(opUpdate.operator, 1);
         }
         updatedOperators[operators[opUpdate.operator].chainID].push(opUpdate);
     }
@@ -149,10 +118,7 @@ contract LagrangeCommittee is
     }
 
     // Returns chain's committee current and next roots at a given block.
-    function getCommittee(
-        uint32 chainID,
-        uint256 blockNumber
-    )
+    function getCommittee(uint32 chainID, uint256 blockNumber)
         public
         view
         returns (CommitteeData memory currentCommittee, uint256 nextRoot)
@@ -164,9 +130,7 @@ contract LagrangeCommittee is
     }
 
     // Computes and returns "next" committee root.
-    function getNext1CommitteeRoot(
-        uint32 chainID
-    ) public view returns (uint256) {
+    function getNext1CommitteeRoot(uint32 chainID) public view returns (uint256) {
         if (committeeLeaves[chainID].length == 0) {
             return _hash2Elements([uint256(0), uint256(0)]);
         } else if (committeeLeaves[chainID].length == 1) {
@@ -231,55 +195,33 @@ contract LagrangeCommittee is
         uint256 nextEpoch = epochNumber + COMMITTEE_NEXT_1;
         committees[chainID][nextEpoch].height = committeeLeaves[chainID].length;
         committees[chainID][nextEpoch].root = nextRoot;
-        committees[chainID][nextEpoch].totalVotingPower = _getTotalVotingPower(
-            chainID
-        );
+        committees[chainID][nextEpoch].totalVotingPower = _getTotalVotingPower(chainID);
     }
 
     // Initializes a new committee, and optionally associates addresses with it.
-    function registerChain(
-        uint32 chainID,
-        uint256 epochPeriod,
-        uint256 freezeDuration
-    ) public onlyOwner {
+    function registerChain(uint32 chainID, uint256 epochPeriod, uint256 freezeDuration) public onlyOwner {
         _initCommittee(chainID, epochPeriod, freezeDuration);
         _update(chainID, 0);
     }
 
     // Adds address stake data and flags it for committee addition
-    function addOperator(
-        address operator,
-        bytes memory blsPubKey,
-        uint32 chainID,
-        uint32 serveUntilBlock
-    ) public onlyService {
+    function addOperator(address operator, bytes memory blsPubKey, uint32 chainID, uint32 serveUntilBlock)
+        public
+        onlyService
+    {
         uint96 stakeAmount = voteWeigher.weightOfOperator(operator, 1);
-        operators[operator] = OperatorStatus(
-            stakeAmount,
-            blsPubKey,
-            serveUntilBlock,
-            chainID,
-            false
-        );
+        operators[operator] = OperatorStatus(stakeAmount, blsPubKey, serveUntilBlock, chainID, false);
     }
 
-    function isUpdatable(
-        uint32 chainID,
-        uint256 epochNumber
-    ) public view returns (bool) {
-        uint256 epochEnd = epochNumber *
-            committeeParams[chainID].duration +
-            committeeParams[chainID].startBlock;
+    function isUpdatable(uint32 chainID, uint256 epochNumber) public view returns (bool) {
+        uint256 epochEnd = epochNumber * committeeParams[chainID].duration + committeeParams[chainID].startBlock;
         uint256 freezeDuration = committeeParams[chainID].freezeDuration;
         return block.number > epochEnd - freezeDuration;
     }
 
     // If applicable, updates committee based on staking, unstaking, and slashing.
     function update(uint32 chainID, uint256 epochNumber) public {
-        require(
-            isUpdatable(chainID, epochNumber),
-            "Block number is prior to committee freeze window."
-        );
+        require(isUpdatable(chainID, epochNumber), "Block number is prior to committee freeze window.");
 
         require(updatedEpoch[chainID] < epochNumber, "Already updated.");
 
@@ -303,10 +245,7 @@ contract LagrangeCommittee is
         updatedEpoch[chainID] = epochNumber;
         delete updatedOperators[chainID];
 
-        emit UpdateCommittee(
-            chainID,
-            bytes32(committees[chainID][epochNumber + COMMITTEE_NEXT_1].root)
-        );
+        emit UpdateCommittee(chainID, bytes32(committees[chainID][epochNumber + COMMITTEE_NEXT_1].root));
     }
 
     function _registerOperator(address operator) internal {
@@ -329,28 +268,21 @@ contract LagrangeCommittee is
         uint256 lastIndex = committeeLeaves[chainID].length - 1;
         address lastAddr = committeeAddrs[chainID][lastIndex];
 
-        committeeLeaves[chainID][leafIndex] = committeeLeaves[chainID][
-            lastIndex
-        ];
+        committeeLeaves[chainID][leafIndex] = committeeLeaves[chainID][lastIndex];
         committeeLeavesMap[chainID][lastAddr] = leafIndex;
         committeeLeaves[chainID].pop();
         committeeAddrs[chainID].pop();
     }
 
     // Computes epoch number for a chain's committee at a given block
-    function getEpochNumber(
-        uint32 chainID,
-        uint256 blockNumber
-    ) public view returns (uint256) {
+    function getEpochNumber(uint32 chainID, uint256 blockNumber) public view returns (uint256) {
         uint256 startBlockNumber = committeeParams[chainID].startBlock;
         uint256 epochPeriod = committeeParams[chainID].duration;
         return (blockNumber - startBlockNumber) / epochPeriod + 1;
     }
 
     // Returns cumulative strategy shares for opted in addresses
-    function _getTotalVotingPower(
-        uint32 chainID
-    ) internal view returns (uint256) {
+    function _getTotalVotingPower(uint32 chainID) internal view returns (uint256) {
         uint256 total = 0;
         for (uint256 i = 0; i < committeeAddrs[chainID].length; i++) {
             total += operators[committeeAddrs[chainID][i]].amount;
@@ -361,50 +293,40 @@ contract LagrangeCommittee is
     function getLeafHash(address opAddr) public view returns (uint256) {
         uint96[11] memory slices;
         OperatorStatus memory opStatus = operators[opAddr];
-        for (uint i = 0; i < 8; i++) {
+        for (uint256 i = 0; i < 8; i++) {
             bytes memory addr = new bytes(12);
-            for (uint j = 0; j < 12; j++) {
+            for (uint256 j = 0; j < 12; j++) {
                 addr[j] = opStatus.blsPubKey[(i * 12) + j];
             }
             bytes12 addrChunk = bytes12(addr);
             slices[i] = uint96(addrChunk);
         }
-        bytes memory addrBytes = abi.encodePacked(
-            opAddr,
-            uint128(opStatus.amount)
-        );
-        for (uint i = 0; i < 3; i++) {
+        bytes memory addrBytes = abi.encodePacked(opAddr, uint128(opStatus.amount));
+        for (uint256 i = 0; i < 3; i++) {
             bytes memory addr = new bytes(12);
-            for (uint j = 0; j < 12; j++) {
+            for (uint256 j = 0; j < 12; j++) {
                 addr[j] = addrBytes[(i * 12) + j];
             }
             bytes12 addrChunk = bytes12(addr);
             slices[i + 8] = uint96(addrChunk);
         }
 
-        return
-            _hash2Elements(
-                [
-                    _hash6Elements(
-                        [
-                            uint256(slices[0]),
-                            uint256(slices[1]),
-                            uint256(slices[2]),
-                            uint256(slices[3]),
-                            uint256(slices[4]),
-                            uint256(slices[5])
-                        ]
-                    ),
-                    _hash5Elements(
-                        [
-                            uint256(slices[6]),
-                            uint256(slices[7]),
-                            uint256(slices[8]),
-                            uint256(slices[9]),
-                            uint256(slices[10])
-                        ]
-                    )
-                ]
-            );
+        return _hash2Elements(
+            [
+                _hash6Elements(
+                    [
+                        uint256(slices[0]),
+                        uint256(slices[1]),
+                        uint256(slices[2]),
+                        uint256(slices[3]),
+                        uint256(slices[4]),
+                        uint256(slices[5])
+                    ]
+                ),
+                _hash5Elements(
+                    [uint256(slices[6]), uint256(slices[7]), uint256(slices[8]), uint256(slices[9]), uint256(slices[10])]
+                )
+            ]
+        );
     }
 }
